Fix binary writer

+++ Master.Benchmarks/Raw/RawSqlite.cs
@@ -49,7 +49,7 @@ internal sealed class RawSqlite : IRawBenchmark
             Type type = array.GetType().GetElementType()?.GetUnderlyingNullableType() ?? throw new ArgumentException(null, nameof(array));
             return type == typeof(int) ? "INT" :
                 type == typeof(string) ? "TEXT" :
-                type == typeof(float) ? "REAL" :
+                type == typeof(double) ? "REAL" :
                 type == typeof(float) ? "REAL" :
                 throw new NotImplementedException();
         }
@@ -60,7 +60,7 @@ internal sealed class RawSqlite : IRawBenchmark
             return type == typeof(int) ? DbType.Int32 :
                 type == typeof(string) ? DbType.String :
                 type == typeof(float) ? DbType.Single :
-                type == typeof(float) ? DbType.Double :
+                type == typeof(double) ? DbType.Double :
                 throw new NotImplementedException();
         }
